updates populi transfer spec
